% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flw_get_contributions.R
\name{flw_get_contributions}
\alias{flw_get_contributions}
\title{Get contribution data from Follow The Money}
\usage{
flw_get_contributions(state = NULL, year = NULL, entity = NULL,
  api_key = NULL)
}
\description{
This package utilizes the Follow The Money API and returns a tidy data frame
of state level political contributions
}
