% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flw_set_api_key.R
\name{flw_set_api_key}
\alias{flw_set_api_key}
\title{Utility function to set API key for the Follow the Money API}
\usage{
flw_set_api_key(api_key)
}
\description{
This function updates the .Renviron file with a new key
}
