% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/followr.R
\name{followr}
\alias{followr}
\title{Access the Follow the Money API}
\description{
This package provides tools to access and download The National Institute on 
Money in State Politics Follow the Money API.  It also provides out data as a 
tidy data frame so that subsequent analysis may be completed using the
\code{tidyverse}(\url{https://tidyverse.org}).
}
